backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-genai
googlemaps
aiohttp
httpx
asyncio
